Made some changes
