--- src/main/java/com/ecom/usersservice/exception/InvalidLoginCredentialsException.java
@@ -1,12 +1,0 @@
-package com.ecom.usersservice.exception;
-
-public class InvalidLoginCredentialsException extends Exception {
-
-	public InvalidLoginCredentialsException() {
-        super();
-    }
-
-    public InvalidLoginCredentialsException(String message) {
-        super(message);
-    } 
-}
