--- src/main/java/com/ecom/usersservice/service/Regex.java
@@ -1,22 +1,0 @@
-package com.ecom.usersservice.service;
-/*
- * package com.easybuy.app.serviceimpl;
- * 
- * import java.util.regex.Matcher; import java.util.regex.Pattern;
- * 
- * public class Regex {
- * 
- * public static boolean isValidPhoneNumber(String phoneNumber) { // This regex
- * matches common formats of phone numbers, but might need // adjustments based
- * on your needs String regex =
- * "^(\\d{10}|\\d{3}-\\d{3}-\\d{4}|\\(\\d{3}\\) \\d{3}-\\d{4})$"; Pattern
- * pattern = Pattern.compile(regex); Matcher matcher =
- * pattern.matcher(phoneNumber); return matcher.matches(); }
- * 
- * public static boolean isValidGmailAddress(String email) { //
- * "^[a-zA-Z0-9._%+-]+@gmail\\.com$"; String regex =
- * "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.com$";
- * 
- * Pattern pattern = Pattern.compile(regex); Matcher matcher =
- * pattern.matcher(email); return matcher.matches(); } }
- */
